Update website metadata and page title generation
This commit updates the metadata in the 'layout.tsx' file, changing the title from "Create Next App" to "Sanity CMS Demo". This is to better reflect the purpose and content of the app. Additionally, a new function 'generateMetadata' was added to 'page.tsx'. This function generates a specific title for each page based on its slug. This improves SEO and provides a better user experience by providing more accurate and informative page titles.