Refactor TextImage component and update usage
The TextImageProps interface was adjusted to use Image and PortableTextBlock types instead of ReactNode, improving specificity and type safety. ImageBox component was introduced to handle image rendering, replacing the previous generic image property. Consequently, changes were applied in Page.tsx to reflect the updated usage of the TextImage component. In HomeHero.tsx, imageUrl was replaced with heroImage in the condition for better clarity of its purpose and usage.