Add the 'sanity-plugin-media' package and update dependencies in the package-lock.json
The 'sanity-plugin-media' package was added as a new package to the project. The addition of the package led to the update of the package-lock.json file with the dependencies installed along with 'sanity-plugin-media'. Alongside this, the dev parameter of some dependencies was also set to true. This addition provides the functionality to the project that the plugin offers.